feat: add MCP server support with FastMCP
