+++ package.json
@@ -17,6 +17,7 @@
     "start": "tsx src/cli/cli.ts",
     "start:api": "tsx src/api/server.ts",
     "start:api:bun": "bun src/api/server.ts",
+    "start:mcp": "tsx src/mcp/server.ts",
     "test": "jest",
     "test:bun": "bun test --exclude tests/web",
     "test:jest:bun": "bun x jest",
@@ -57,6 +58,7 @@
     "@hono/zod-openapi": "^1.3.0",
     "commander": "^14.0.3",
     "eastasianwidth": "^0.3.0",
+    "fastmcp": "^4.0.1",
     "glob": "^13.0.6",
     "hono": "^4.12.18",
     "sanitize-html": "^2.17.3",
